tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-html>=3.2.0
boto3>=1.26.0
botocore>=1.29.0
moto[all]>=5.0.0
requests>=2.28.0
aws-cdk-lib>=2.80.0
constructs>=10.0.0
